pyproject.toml
[project]
name = "doco"
version = "1.0.0"
description = "Document generation service with MCP server support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.100.0",
    "starlette>=0.27.0",
    "sse-starlette>=1.6.0",
    "mcp>=0.9.0",
    "PyJWT>=2.8.0",
    "httpx>=0.24.0",
    "Jinja2>=3.1.0",
    "WeasyPrint>=60.0",
    "html2text>=2020.1.16",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.uv.dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
filterwarnings = [
    "ignore::RuntimeWarning:matplotlib",
]
